fix: move related parts into hover_links and document_symbols
