+++ crates/editor/src/navigation.rs
@@ -853,6 +853,17 @@ impl Editor {
         });
     }
 
+    pub fn select_to_end(&mut self, _: &SelectToEnd, window: &mut Window, cx: &mut Context<Self>) {
+        let buffer = self.buffer.read(cx).snapshot(cx);
+        let mut selection = self
+            .selections
+            .first::<MultiBufferOffset>(&self.display_snapshot(cx));
+        selection.set_head(buffer.len(), SelectionGoal::None);
+        self.change_selections(Default::default(), window, cx, |s| {
+            s.select(vec![selection]);
+        });
+    }
+
     pub fn move_to_end(&mut self, _: &MoveToEnd, window: &mut Window, cx: &mut Context<Self>) {
         if matches!(self.mode, EditorMode::SingleLine) {
             cx.propagate();
@@ -1075,18 +1086,6 @@ impl Editor {
         cx: &mut Context<Self>,
     ) {
         self.go_to_singleton_buffer_range(point..point, window, cx);
-    }
-
-    /// Like `go_to_singleton_buffer_point`, but does not push a navigation
-    /// history entry. Useful when the caller already recorded one (e.g. when
-    /// a file was just opened and we only need to move the cursor).
-    pub fn go_to_singleton_buffer_point_silently(
-        &mut self,
-        point: Point,
-        window: &mut Window,
-        cx: &mut Context<Self>,
-    ) {
-        self.go_to_singleton_buffer_range_impl(point..point, false, window, cx);
     }
 
     pub(super) fn go_to_next_change(
@@ -1316,86 +1315,6 @@ impl Editor {
                 .await?;
             anyhow::Ok(navigated)
         })
-    }
-
-    pub fn open_url(&mut self, _: &OpenUrl, window: &mut Window, cx: &mut Context<Self>) {
-        let selection = self.selections.newest_anchor();
-        let head = selection.head();
-        let tail = selection.tail();
-
-        let Some((buffer, start_position)) =
-            self.buffer.read(cx).text_anchor_for_position(head, cx)
-        else {
-            return;
-        };
-
-        let end_position = if head != tail {
-            let Some((_, pos)) = self.buffer.read(cx).text_anchor_for_position(tail, cx) else {
-                return;
-            };
-            Some(pos)
-        } else {
-            None
-        };
-
-        let url_finder = cx.spawn_in(window, async move |_editor, cx| {
-            let url = if let Some(end_pos) = end_position {
-                find_url_from_range(&buffer, start_position..end_pos, cx.clone())
-            } else {
-                find_url(&buffer, start_position, cx.clone()).map(|(_, url)| url)
-            };
-
-            if let Some(url) = url {
-                cx.update(|window, cx| {
-                    if parse_zed_link(&url, cx).is_some() {
-                        window.dispatch_action(Box::new(zed_actions::OpenZedUrl { url }), cx);
-                    } else {
-                        cx.open_url(&url);
-                    }
-                })?;
-            }
-
-            anyhow::Ok(())
-        });
-
-        url_finder.detach();
-    }
-
-    pub fn open_selected_filename(
-        &mut self,
-        _: &OpenSelectedFilename,
-        window: &mut Window,
-        cx: &mut Context<Self>,
-    ) {
-        let Some(workspace) = self.workspace() else {
-            return;
-        };
-
-        let position = self.selections.newest_anchor().head();
-
-        let Some((buffer, buffer_position)) =
-            self.buffer.read(cx).text_anchor_for_position(position, cx)
-        else {
-            return;
-        };
-
-        let project = self.project.clone();
-
-        cx.spawn_in(window, async move |_, cx| {
-            let result = find_file(&buffer, project, buffer_position, cx).await;
-
-            if let Some((_, file_target)) = result {
-                let item = workspace
-                    .update_in(cx, |workspace, window, cx| {
-                        workspace.open_resolved_path(file_target.resolved_path.clone(), window, cx)
-                    })?
-                    .await?;
-
-                file_target.navigate_item_to_position(item, cx);
-            }
-            anyhow::Ok(())
-        })
-        .detach();
     }
 
     pub(super) fn navigate_to_hover_links(
@@ -1788,114 +1707,6 @@ impl Editor {
         if let Some(task) = task {
             task.detach();
         };
-    }
-
-    pub(super) fn go_to_symbol_by_offset(
-        &mut self,
-        window: &mut Window,
-        cx: &mut Context<Self>,
-        offset: i8,
-    ) -> Task<Result<()>> {
-        let editor_snapshot = self.snapshot(window, cx);
-
-        // We don't care about multi-buffer symbols
-        if !editor_snapshot.is_singleton() {
-            return Task::ready(Ok(()));
-        }
-
-        let cursor_offset = self
-            .selections
-            .newest::<MultiBufferOffset>(&editor_snapshot.display_snapshot)
-            .head();
-
-        cx.spawn_in(window, async move |editor, wcx| -> Result<()> {
-            let Ok(Some(remote_id)) = editor.update(wcx, |ed, cx| {
-                let buffer = ed.buffer.read(cx).as_singleton()?;
-                Some(buffer.read(cx).remote_id())
-            }) else {
-                return Ok(());
-            };
-
-            let task = editor.update(wcx, |ed, cx| ed.buffer_outline_items(remote_id, cx))?;
-            let outline_items: Vec<OutlineItem<text::Anchor>> = task.await;
-
-            let multi_snapshot = editor_snapshot.buffer();
-            let buffer_range = |range: &Range<_>| {
-                Some(
-                    multi_snapshot
-                        .buffer_anchor_range_to_anchor_range(range.clone())?
-                        .to_offset(multi_snapshot),
-                )
-            };
-
-            wcx.update_window(wcx.window_handle(), |_, window, acx| {
-                let current_idx = outline_items
-                    .iter()
-                    .enumerate()
-                    .filter_map(|(idx, item)| {
-                        // Find the closest outline item by distance between outline text and cursor location
-                        let source_range = buffer_range(&item.source_range_for_text)?;
-                        let distance_to_closest_endpoint = cmp::min(
-                            (source_range.start.0 as isize - cursor_offset.0 as isize).abs(),
-                            (source_range.end.0 as isize - cursor_offset.0 as isize).abs(),
-                        );
-
-                        let item_towards_offset =
-                            (source_range.start.0 as isize - cursor_offset.0 as isize).signum()
-                                == (offset as isize).signum();
-
-                        let source_range_contains_cursor = source_range.contains(&cursor_offset);
-
-                        // To pick the next outline to jump to, we should jump in the direction of the offset, and
-                        // we should not already be within the outline's source range. We then pick the closest outline
-                        // item.
-                        (item_towards_offset && !source_range_contains_cursor)
-                            .then_some((distance_to_closest_endpoint, idx))
-                    })
-                    .min()
-                    .map(|(_, idx)| idx);
-
-                let Some(idx) = current_idx else {
-                    return;
-                };
-
-                let Some(range) = buffer_range(&outline_items[idx].source_range_for_text) else {
-                    return;
-                };
-                let selection = [range.start..range.start];
-
-                let _ = editor
-                    .update(acx, |editor, ecx| {
-                        editor.change_selections(
-                            SelectionEffects::scroll(Autoscroll::newest()),
-                            window,
-                            ecx,
-                            |s| s.select_ranges(selection),
-                        );
-                    })
-                    .ok();
-            })?;
-
-            Ok(())
-        })
-    }
-
-    pub(super) fn go_to_next_symbol(
-        &mut self,
-        _: &GoToNextSymbol,
-        window: &mut Window,
-        cx: &mut Context<Self>,
-    ) {
-        self.go_to_symbol_by_offset(window, cx, 1).detach();
-    }
-
-    pub(super) fn go_to_previous_symbol(
-        &mut self,
-        _: &GoToPreviousSymbol,
-        window: &mut Window,
-        cx: &mut Context<Self>,
-    ) {
-        self.go_to_symbol_by_offset(window, cx, -1).detach();
     }
 
     pub fn go_to_reference_before_or_after_position(
@@ -2306,6 +2117,18 @@ impl Editor {
         cx: &mut Context<Self>,
     ) {
         self.go_to_singleton_buffer_range_impl(range, true, window, cx);
+    }
+
+    /// Like `go_to_singleton_buffer_point`, but does not push a navigation
+    /// history entry. Useful when the caller already recorded one (e.g. when
+    /// a file was just opened and we only need to move the cursor).
+    pub(super) fn go_to_singleton_buffer_point_silently(
+        &mut self,
+        point: Point,
+        window: &mut Window,
+        cx: &mut Context<Self>,
+    ) {
+        self.go_to_singleton_buffer_range_impl(point..point, false, window, cx);
     }
 
     fn go_to_singleton_buffer_range_impl(
